+++ NotWuphfAPI/NotWuphfAPI.Web/Controllers/GroupsController.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using NotWuphfAPI.Core.Auth.Model;
 using NotWuphfAPI.Core.DTO;
 using NotWuphfAPI.Core.Entities;
 using NotWuphfAPI.Core.Extensions;
@@ -14,10 +17,12 @@ namespace NotWuphfAPI.Web.Controllers
     public class GroupsController : ControllerBase
     {
         private readonly IRepository<Group> _groupsRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public GroupsController(IRepository<Group> groupsRepository)
+        public GroupsController(IRepository<Group> groupsRepository, IAuthorizationService authorizationService)
         {
             _groupsRepository = groupsRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -42,6 +47,7 @@ namespace NotWuphfAPI.Web.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = GroupRoles.GroupUser)]
         public async Task<ActionResult<GroupDTO>> Create(CreateGroupDTO createGroupDTO)
         {
             var group = new Group()
@@ -50,7 +56,10 @@ namespace NotWuphfAPI.Web.Controllers
                 Description = createGroupDTO.Description,
                 Visibility = createGroupDTO.Visibility,
                 CreationDate = DateTime.UtcNow,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
             };
+            
+            Console.WriteLine(User);
 
             await _groupsRepository.AddAsync(group);
 
@@ -58,12 +67,20 @@ namespace NotWuphfAPI.Web.Controllers
         }
 
         [HttpPut("{groupId}")]
+        [Authorize(Roles = GroupRoles.GroupUser)]
         public async Task<ActionResult<GroupDTO>> Update(int groupId, UpdateGroupDTO updateGroupDTO)
         {
             var spec = new GroupByIdSpec(groupId);
             var group = await _groupsRepository.FirstOrDefaultAsync(spec);
 
             if (group is null) return NotFound();
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, group, PolicyNames.ResourceOwner);
+            
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             group.Name = updateGroupDTO.Name;
             group.Description = updateGroupDTO.Description;
@@ -75,12 +92,20 @@ namespace NotWuphfAPI.Web.Controllers
         }
 
         [HttpDelete("{groupId}")]
+        [Authorize(Roles = GroupRoles.GroupUser)]
         public async Task<ActionResult> Remove(int groupId)
         {
             var spec = new GroupByIdSpec(groupId);
             var group = await _groupsRepository.FirstOrDefaultAsync(spec);
 
             if (group is null) return NotFound();
+            
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, group, PolicyNames.ResourceOwner);
+            
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             await _groupsRepository.DeleteAsync(group);
